requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pillow
python-multipart
requests
python-dotenv
